requirements.txt
ultralytics
opencv-python
pandas
streamlit
pyyaml
requests
numpy
torch
torchvision
torchaudio
